+++ schemas/ispyb/routines.sql
@@ -7794,6 +7794,39 @@ DELIMITER ;
 /*!50003 SET character_set_results = @saved_cs_results */ ;
 /*!50003 SET collation_connection  = @saved_col_connection */ ;
 /*!50003 SET @saved_sql_mode       = @@sql_mode */ ;
+/*!50003 SET sql_mode              = 'STRICT_TRANS_TABLES,ERROR_FOR_DIVISION_BY_ZERO,NO_AUTO_CREATE_USER,NO_ENGINE_SUBSTITUTION' */ ;
+/*!50003 DROP PROCEDURE IF EXISTS `update_dc_append_comments` */;
+/*!50003 SET @saved_cs_client      = @@character_set_client */ ;
+/*!50003 SET @saved_cs_results     = @@character_set_results */ ;
+/*!50003 SET @saved_col_connection = @@collation_connection */ ;
+/*!50003 SET character_set_client  = utf8mb3 */ ;
+/*!50003 SET character_set_results = utf8mb3 */ ;
+/*!50003 SET collation_connection  = utf8mb3_general_ci */ ;
+DELIMITER ;;
+CREATE PROCEDURE `update_dc_append_comments`(
+  p_id int(11) unsigned,
+  p_comments varchar(1024),
+  p_separator varchar(5)
+)
+    MODIFIES SQL DATA
+    COMMENT 'Appends text to DataCollection.comments for dataCollectionId=p_id. '
+BEGIN
+  IF NOT (p_id IS NULL) AND NOT (p_comments IS NULL) THEN
+
+    UPDATE DataCollection 
+    SET comments = concat_ws(p_separator, comments, p_comments) 
+    WHERE dataCollectionId = p_id; 
+
+  ELSE
+      SIGNAL SQLSTATE '45000' SET MYSQL_ERRNO=1644, MESSAGE_TEXT='Mandatory arguments p_id and/or p_comments are NULL';
+  END IF;
+END ;;
+DELIMITER ;
+/*!50003 SET sql_mode              = @saved_sql_mode */ ;
+/*!50003 SET character_set_client  = @saved_cs_client */ ;
+/*!50003 SET character_set_results = @saved_cs_results */ ;
+/*!50003 SET collation_connection  = @saved_col_connection */ ;
+/*!50003 SET @saved_sql_mode       = @@sql_mode */ ;
 /*!50003 SET sql_mode              = 'STRICT_TRANS_TABLES' */ ;
 /*!50003 DROP PROCEDURE IF EXISTS `update_dc_experiment` */;
 /*!50003 SET @saved_cs_client      = @@character_set_client */ ;